+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# (next major)
+
+- Removed RBWSMethod.unrecognized
+
 ## 1.1.4
 
 - Deprecated RBWSMethod.unrecognized, will be removed in next major release.